fix des pbm de l'auth

+++ api/src/Controller/AuthController.php
@@ -36,6 +36,7 @@ class AuthController extends AbstractController
         $username = $data['username'] ?? null;
         $password = $data['password'] ?? null;
         $locale = $data['locale'] ?? null;
+        $rememberMe = $data['rememberMe'] ?? false;
 
         if (!$username || !$password) {
             return new JsonResponse(
@@ -94,7 +95,8 @@ class AuthController extends AbstractController
         }
 
         $token = $this->jwtManager->create($user);
-        $refreshToken = $this->refreshTokenGenerator->createForUserWithTtl($user, 86400 * 7); // 7 days
+        $refreshTokenTtl = $rememberMe ? 86400 * 30 : 3600; // 30 days or 1 hour
+        $refreshToken = $this->refreshTokenGenerator->createForUserWithTtl($user, $refreshTokenTtl);
 
         $refreshTokenEntity = $this->entityManager->getRepository(RefreshToken::class)->findOneBy(['username' => $user->getEmail()]);
 
@@ -116,7 +118,7 @@ class AuthController extends AbstractController
         $response->headers->setCookie(
             Cookie::create('refresh_token')
                 ->withValue($refreshToken->getRefreshToken())
-                ->withExpires(strtotime('+1 days'))
+                ->withExpires(strtotime($rememberMe ? '+30 days' : '+1 hour'))
                 ->withPath('/')
                 ->withSecure(false) // En local, Secure doit être désactivé (Activer en prod)
                 ->withHttpOnly(true)
@@ -166,13 +168,17 @@ class AuthController extends AbstractController
         }
 
         $newToken = $this->jwtManager->create($user);
-        $newRefreshToken = new RefreshToken();
-        $newRefreshToken->setRefreshToken()
-            ->setUsername($user->getEmail())
-            ->setValid((new \DateTime())->modify('+1 week'));
+        if ($refreshToken->getValid() > new \DateTime()) {
+            $newRefreshToken = $refreshToken;
+        } else {
+            $newRefreshToken = new RefreshToken();
+            $newRefreshToken->setRefreshToken()
+                ->setUsername($user->getEmail())
+                ->setValid((new \DateTime())->modify('+1 hour'));
 
-        $refreshTokenManager->delete($refreshToken);
-        $refreshTokenManager->save($newRefreshToken);
+            $refreshTokenManager->delete($refreshToken);
+            $refreshTokenManager->save($newRefreshToken);
+        }
 
         $response = new JsonResponse(['token' => $newToken]);
 
@@ -195,7 +201,7 @@ class AuthController extends AbstractController
         $data = json_decode($request->getContent(), true);
         $userId = $data['user_id'] ?? null;
         $otpCode = $data['code'] ?? null;
-
+        $rememberMe = $data['rememberMe'] ?? false;
 
         if (!$userId || !$otpCode) {
             return new JsonResponse(['message' => 'Missing parameters'], JsonResponse::HTTP_BAD_REQUEST);
@@ -217,12 +223,18 @@ class AuthController extends AbstractController
         $this->entityManager->flush();
 
         $token = $this->jwtManager->create($user);
-        $refreshToken = $this->refreshTokenGenerator->createForUserWithTtl($user, 86400 * 7);
+        $refreshTokenTtl = $rememberMe ? 86400 * 30 : 3600; // 30 days or 1 hour
+        $refreshToken = $this->refreshTokenGenerator->createForUserWithTtl($user, $refreshTokenTtl);
 
-        $refreshTokenEntity = new RefreshToken();
+        $refreshTokenEntity = $this->entityManager->getRepository(RefreshToken::class)->findOneBy(['username' => $user->getEmail()]);
+
+        if (!$refreshTokenEntity) {
+            $refreshTokenEntity = new RefreshToken();
+        }
+
         $refreshTokenEntity->setRefreshToken($refreshToken->getRefreshToken())
-        ->setUsername($user->getEmail())
-        ->setValid($refreshToken->getValid());
+            ->setUsername($user->getEmail())
+            ->setValid($refreshToken->getValid());
 
         $this->entityManager->persist($refreshTokenEntity);
         $this->entityManager->flush();
@@ -232,7 +244,7 @@ class AuthController extends AbstractController
         $response->headers->setCookie(
             Cookie::create('refresh_token')
             ->withValue($refreshToken->getRefreshToken())
-            ->withExpires(strtotime('+1 days'))
+            ->withExpires(strtotime($rememberMe ? '+30 days' : '+1 hour'))
             ->withPath('/')
             ->withSecure(false) // En local, Secure doit être désactivé (Activer en prod)
             ->withHttpOnly(true)
